Moving functions to separate modules with single responsibility
